==> categorical_attrition/scripts/02_sample.R <==

# Script for sampling coefficients and variance-covariance matrices from a
# multivariate normal distribution and producing datasets for plotting
# Outputs are not saved in github repo as file sizes are quite large, but
# the script can be run using uploaded data to get the outputs

source("config.R")
source("categorical_attrition/scripts/00_functions.R")

library(tidyverse)
library(MASS)
library(flexsurv)

# Import prepared data
lst_prep <- read_rds("categorical_attrition/processed_data/prepared_data.rds")

# Check empirical cumulative incidence curves
plot_cind <- unnest(lst_prep$cind, cols = "cind") %>% 
  ggplot(aes(x = time, y = 1 - estimate, colour = cause)) +
  geom_step(linewidth = 1) +
  facet_wrap(~ctgov, scales = "free") +
  labs(x = "Time in trial (days)", y = "Cumulative incidence of attrition")

plot_cind

# Sample parameters from a multivariate distribution for each trial model
sample_parameters <- get_samples()

# Estimate cumulative incidence using sampled parameters
cind_sampled_estimates <- get_cind()

# Merge cumulative incidence datasets
cind_bind <- lapply(seq_along(cind_sampled_estimates), function(i) {
  name <- names(cind_sampled_estimates[i])
  name_parts <- strsplit(name, "_")[[1]]
  name1 <- name_parts[1]
  name2 <- name_parts[2]
  lapply(seq_along(cind_sampled_estimates[[i]]), function(j) {
    name3 <- names(cind_sampled_estimates[[i]][j])
    bind_rows(cind_sampled_estimates[[i]][[j]]) %>% 
      mutate(
        ctgov = name1,
        cause = name2,
        dist = name3
      )
  })
})

cind_bind <- bind_rows(cind_bind) %>% distinct()
# saveRDS(cind_bind, "categorical_attrition/processed_data/cind_sampled_df.rds")

# Estimate hazard rates over time using model parameters
haz_bind <- get_hazards()

# Save (locally, can run the script as said above to get these objects)
# saveRDS(haz_bind, "categorical_attrition/processed_data/haz_sampled_df.rds")

# Unload MASS after use to avoid dplyr conflicts
detach("package:MASS", unload = TRUE)
